AT-12 reorganized layout
